# Conflicts:
#	sormas-api/src/main/resources/captions_de-CH.properties
#	sormas-api/src/main/resources/captions_de-DE.properties
#	sormas-api/src/main/resources/captions_en-AF.properties
#	sormas-api/src/main/resources/captions_es-CU.properties
#	sormas-api/src/main/resources/captions_es-EC.properties
#	sormas-api/src/main/resources/captions_es-ES.properties
#	sormas-api/src/main/resources/captions_fa-AF.properties
#	sormas-api/src/main/resources/captions_fi-FI.properties
#	sormas-api/src/main/resources/captions_fil-PH.properties
#	sormas-api/src/main/resources/captions_fj-FJ.properties
#	sormas-api/src/main/resources/captions_fr-CH.properties
#	sormas-api/src/main/resources/captions_fr-FR.properties
#	sormas-api/src/main/resources/captions_haw-US.properties
#	sormas-api/src/main/resources/captions_hi-IN.properties
#	sormas-api/src/main/resources/captions_it-CH.properties
#	sormas-api/src/main/resources/captions_it-IT.properties
#	sormas-api/src/main/resources/captions_ja-JP.properties
#	sormas-api/src/main/resources/captions_ks-PK.properties
#	sormas-api/src/main/resources/captions_ku-TR.properties
#	sormas-api/src/main/resources/captions_la-LA.properties
#	sormas-api/src/main/resources/captions_mt-MT.properties
#	sormas-api/src/main/resources/captions_nl-NL.properties
#	sormas-api/src/main/resources/captions_no-NO.properties
#	sormas-api/src/main/resources/captions_pl-PL.properties
#	sormas-api/src/main/resources/captions_ps-AF.properties
#	sormas-api/src/main/resources/captions_pt-PT.properties
#	sormas-api/src/main/resources/captions_ro-RO.properties
#	sormas-api/src/main/resources/captions_ru-RU.properties
#	sormas-api/src/main/resources/captions_sv-SE.properties
#	sormas-api/src/main/resources/captions_sw-KE.properties
#	sormas-api/src/main/resources/captions_tr-TR.properties
#	sormas-api/src/main/resources/captions_uk-UA.properties
#	sormas-api/src/main/resources/captions_ur-PK.properties
#	sormas-api/src/main/resources/captions_zh-CN.properties
#	sormas-api/src/main/resources/captions_zu-ZA.properties
#	sormas-api/src/main/resources/strings_es-EC.properties
#	sormas-api/src/main/resources/strings_es-ES.properties
#	sormas-api/src/main/resources/strings_fi-FI.properties
#	sormas-api/src/main/resources/strings_fr-FR.properties
#	sormas-api/src/main/resources/strings_it-IT.properties

+++ sormas-api/src/main/java/de/symeda/sormas/api/event/EventExportDto.java
@@ -29,7 +29,7 @@ public class EventExportDto implements Serializable {
 	private String uuid;
 	private String externalId;
 	private EventStatus eventStatus;
-	private int participantsCount;
+	private long participantsCount;
 	private Disease disease;
 	private String diseaseDetails;
 	private Date startDate;
@@ -59,7 +59,7 @@ public class EventExportDto implements Serializable {
 		String uuid,
 		String externalId,
 		EventStatus eventStatus,
-		int participantsCount,
+		long participantsCount,
 		Disease disease,
 		String diseaseDetails,
 		Date startDate,
@@ -310,11 +310,11 @@ public class EventExportDto implements Serializable {
 	}
 
 	@Order(25)
-	public int getParticipantsCount() {
+	public long getParticipantsCount() {
 		return participantsCount;
 	}
 
-	public void setParticipantsCount(int participantsCount) {
+	public void setParticipantsCount(long participantsCount) {
 		this.participantsCount = participantsCount;
 	}
